Security: cast base64-decoded ID parameters to int (SQL injection fix)
Task/writer IDs arrive as base64-encoded GET parameters and were
interpolated into SQL as raw decoded strings, allowing injection via
a crafted parameter. All IDs are numeric row ids, so casting at the
decode site closes the hole in every downstream query without
changing behavior. Matches the (int) idiom newer files already use.

+++ confirmation.php
@@ -152,7 +152,7 @@ function sendEmail($writer, $pages, $cpp, $due_date, $writerEmail, $taskId, $act
 
 if (isset($_GET['task_id']) && isset($_GET['action'])) {
     $encodedId = $_GET['task_id'];
-    $taskId = base64_decode($encodedId);
+    $taskId = (int) base64_decode($encodedId);
     $action = $_GET['action'];
 
     // Fetch task details
