+++ sudo/overdraft.php
@@ -103,7 +103,7 @@ if (isset($_POST['save'])) {
 
 if (isset($_GET['delete'])) {
     $encodedId = $_GET['delete'];
-    $cmpid = base64_decode($encodedId);
+    $cmpid = (int) base64_decode($encodedId);
 
     // Validate $cmpid to ensure it's numeric and not empty
     if (is_numeric($cmpid) && !empty($cmpid)) {
