+++ sudo/tasks-in-progress.php
@@ -15,7 +15,7 @@ $msg = "";
 
 if (isset($_GET['del'])) {
     $encodedId = $_GET['del'];
-    $cmpid = base64_decode($encodedId);
+    $cmpid = (int) base64_decode($encodedId);
 
     // Validate $cmpid to ensure it's numeric and not empty
     if (is_numeric($cmpid) && !empty($cmpid)) {
